pyproject.toml
[tool.poetry]
name = "repl-nix-outfitwizard"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.38.0"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.2.0"
numpy = "^2.1.1"
scikit-learn = "^1.5.2"
scipy = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
